+++ test/tests/timer.test.js
@@ -23,17 +23,45 @@
 
 import {PicoGL} from "../../src/picogl.js";
 
-glCheck("PicoGL", (t) => {
-    t.ok(PicoGL, "PicoGL was loaded");
-    t.ok(typeof PicoGL.FLOAT === "number", "PicoGL.FLOAT is a number");
-    t.ok(typeof PicoGL.COMPRESSED_SRGB_S3TC_DXT1_EXT === "number", "PicoGL.COMPRESSED_SRGB_S3TC_DXT1_EXT is a number");
+glCheck("Timer CPU timing", async (t, canvas) => {
+    let app = PicoGL.createApp(canvas);
+    let timer = app.createTimer();
+
+    t.ok(timer.gl, "Timer contains a gl context");
+    t.ok(timer.cpuTimer, "Timer created a cpu timer");
+
+    timer.start();
+    timer.end();
+
+    await t.loopUntil(() => timer.ready());
+
+    t.equal(typeof timer.cpuTime, "number", "CPU time is number");
     t.done();
+
 });
 
-glCheck("PicoGL.createApp", (t, canvas) => {
-    let app = PicoGL.createApp(canvas);
+glCheck("Timer GPU timing", async (t, canvas) => {
+    if (!PicoGL.WEBGL_INFO.GPU_TIMER) {
+        t.ok(true, "GPU Timing not supported");
+        t.done();
+        return;
+    }
 
-    t.ok(app, "App was created");
-    t.ok(typeof PicoGL.WEBGL_INFO.MAX_TEXTURE_UNITS === "number", "WEBGL_INFO initialized");
+    let app = PicoGL.createApp(canvas);
+    let timer = app.createTimer();
+
+    t.ok(timer.gl, "Timer contains a gl context");
+    t.ok(timer.gpuTimerQuery, "Timer created a gpu timer");
+
+    timer.start();
+    timer.end();
+
+    await t.loopUntil(() => timer.ready()); 
+        
+    t.equal(typeof timer.gpuTime, "number", "GPU time is number");
+
+    timer.delete();
+    t.equal(timer.gpuTimerQuery, null, "Timer was deleted");
+
     t.done();
 });