+++ test/tests/utils.js
@@ -21,19 +21,50 @@
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 ///////////////////////////////////////////////////////////////////////////////////
 
-import {PicoGL} from "../../src/picogl.js";
+export function createQuadDrawCall(app, fs) {
+    let vertexArray = app.createVertexArray()
+        .vertexAttributeBuffer(0, 
+            app.createVertexBuffer(app.gl.FLOAT, 2, new Float32Array([
+                -1, -1,
+                1, -1,
+                -1, 1,
+                1, 1
+            ]))
+        );
+    let program = app.createProgram(`
+        #version 300 es
+        
+        layout(location=0) in vec4 position;
+        
+        out vec2 vUV;
+        void main() {
+            vUV = position.xy * 0.5 + 0.5;
+            gl_Position = position;
+        }
+    `,
+    fs);
 
-glCheck("PicoGL", (t) => {
-    t.ok(PicoGL, "PicoGL was loaded");
-    t.ok(typeof PicoGL.FLOAT === "number", "PicoGL.FLOAT is a number");
-    t.ok(typeof PicoGL.COMPRESSED_SRGB_S3TC_DXT1_EXT === "number", "PicoGL.COMPRESSED_SRGB_S3TC_DXT1_EXT is a number");
-    t.done();
-});
+    app.cullBackfaces();
 
-glCheck("PicoGL.createApp", (t, canvas) => {
-    let app = PicoGL.createApp(canvas);
+    return app.createDrawCall(program, vertexArray).primitive(app.gl.TRIANGLE_STRIP);
+}
 
-    t.ok(app, "App was created");
-    t.ok(typeof PicoGL.WEBGL_INFO.MAX_TEXTURE_UNITS === "number", "WEBGL_INFO initialized");
-    t.done();
-});
+export function loadImages(urls) {
+    return new Promise((resolve) => {
+        let numImages = urls.length;
+
+        let images = new Array(numImages);
+
+        function onload() {
+            if (--numImages === 0) {
+                resolve(images);
+            }
+        }
+
+        for (let i = 0; i < numImages; ++i) {
+            images[i] = new Image();
+            images[i].onload = onload;
+            images[i].src = urls[i];
+        }
+    });
+}